+++ src/test/java/org/opensearch/knn/index/AdvancedFilteringUseCasesIT.java
@@ -43,8 +43,6 @@ import static org.opensearch.knn.common.KNNConstants.TYPE_KNN_VECTOR;
 import static org.opensearch.knn.common.KNNConstants.TYPE_NESTED;
 import static org.opensearch.knn.common.KNNConstants.VECTOR;
 
-import static org.hamcrest.Matchers.containsString;
-
 /**
  * This class contains the IT for some advanced and tricky use-case of filters.
  * <a href="https://github.com/opensearch-project/k-NN/issues/1356">Github issue</a>
@@ -193,7 +191,7 @@ public class AdvancedFilteringUseCasesIT extends KNNRestTestCase {
      */
     @SneakyThrows
     @ExpectRemoteBuildValidation
-    public void testFiltering_whenNestedKNNAndFilterFieldWithNoNestedContextInFilterQuery_thenFailure() {
+    public void testFiltering_whenNestedKNNAndFilterFieldWithNoNestedContextInFilterQuery_thenSuccess() {
         for (final String engine : enginesToTest) {
             // Set up the index with nested k-nn and metadata fields
             createKnnIndex(INDEX_NAME, createNestedMappings(1, engine));
@@ -223,8 +221,7 @@ public class AdvancedFilteringUseCasesIT extends KNNRestTestCase {
             builder.endObject();
             builder.endObject().endObject().endObject().endObject().endObject().endObject();
 
-            final AssertionError error = expectThrows(AssertionError.class, () -> validateFilterSearch(builder.toString(), engine));
-            assertThat(error.getMessage(), containsString("expected:<" + DOCUMENT_IN_RESPONSE + "> but was:<0>"));
+            validateFilterSearch(builder.toString(), engine);
 
             // cleanup
             deleteKNNIndex(INDEX_NAME);